Created a utils file
